+++ portal/tests/Integration/LogRetentionTest.php
@@ -16,6 +16,7 @@ class LogRetentionTest extends TestCase
     private const TYPE    = 'phpunit_retention';
     private const ACTION  = 'phpunit_retention';
 
+    #[\Override]
     protected function tearDown(): void
     {
         db()->prepare('DELETE FROM error_log WHERE channel = ?')->execute([self::CHANNEL]);
@@ -27,10 +28,10 @@ class LogRetentionTest extends TestCase
     {
         db()->prepare(
             "INSERT INTO error_log (logged_at, level, channel, message) VALUES (?, 'info', ?, 'old')"
-        )->execute([date('Y-m-d H:i:s', strtotime('-2 months')), self::CHANNEL]);
+        )->execute([date('Y-m-d H:i:s', (int) strtotime('-2 months')), self::CHANNEL]);
         db()->prepare(
             "INSERT INTO error_log (logged_at, level, channel, message) VALUES (?, 'info', ?, 'new')"
-        )->execute([date('Y-m-d H:i:s', strtotime('-1 day')), self::CHANNEL]);
+        )->execute([date('Y-m-d H:i:s', (int) strtotime('-1 day')), self::CHANNEL]);
 
         apply_log_retention();
 
@@ -45,10 +46,10 @@ class LogRetentionTest extends TestCase
     {
         db()->prepare(
             "INSERT INTO email_log (sent_at, to_email, subject, type, status) VALUES (?, 'x@test.invalid', 'old', ?, 'sent')"
-        )->execute([date('Y-m-d H:i:s', strtotime('-7 months')), self::TYPE]);
+        )->execute([date('Y-m-d H:i:s', (int) strtotime('-7 months')), self::TYPE]);
         db()->prepare(
             "INSERT INTO email_log (sent_at, to_email, subject, type, status) VALUES (?, 'x@test.invalid', 'new', ?, 'sent')"
-        )->execute([date('Y-m-d H:i:s', strtotime('-1 day')), self::TYPE]);
+        )->execute([date('Y-m-d H:i:s', (int) strtotime('-1 day')), self::TYPE]);
 
         apply_log_retention();
 
@@ -63,10 +64,10 @@ class LogRetentionTest extends TestCase
     {
         db()->prepare(
             "INSERT INTO activity_log (action, created_at) VALUES (?, ?)"
-        )->execute([self::ACTION, date('Y-m-d H:i:s', strtotime('-7 months'))]);
+        )->execute([self::ACTION, date('Y-m-d H:i:s', (int) strtotime('-7 months'))]);
         db()->prepare(
             "INSERT INTO activity_log (action, created_at) VALUES (?, ?)"
-        )->execute([self::ACTION, date('Y-m-d H:i:s', strtotime('-1 day'))]);
+        )->execute([self::ACTION, date('Y-m-d H:i:s', (int) strtotime('-1 day'))]);
 
         apply_log_retention();
 
@@ -81,7 +82,7 @@ class LogRetentionTest extends TestCase
         // 3 weeks old — inside the 1-month error_log window, must survive.
         db()->prepare(
             "INSERT INTO error_log (logged_at, level, channel, message) VALUES (?, 'info', ?, 'recent')"
-        )->execute([date('Y-m-d H:i:s', strtotime('-3 weeks')), self::CHANNEL]);
+        )->execute([date('Y-m-d H:i:s', (int) strtotime('-3 weeks')), self::CHANNEL]);
 
         apply_log_retention();
